move tests to str/tests
