+++ src/lib.rs
@@ -108,3 +108,5 @@ pub const DCC_MIME_DEBUG: &str = "DCC_MIME_DEBUG";
 
 #[cfg(test)]
 mod test_utils;
+#[cfg(test)]
+mod tests;
